+++ lib/Service/AccountService.php
@@ -7,6 +7,7 @@ use OCA\Libresign\AppInfo\Application;
 use OCA\Libresign\Db\FileMapper;
 use OCA\Libresign\Db\FileUser;
 use OCA\Libresign\Db\FileUserMapper;
+use OCA\Libresign\Db\ReportDao;
 use OCA\Libresign\Exception\LibresignException;
 use OCA\Libresign\Handler\CfsslHandler;
 use OCA\Libresign\Helper\JSActions;
@@ -14,10 +15,13 @@ use OCA\Settings\Mailer\NewUserMailHelper;
 use OCP\Files\File;
 use OCP\Files\IRootFolder;
 use OCP\IConfig;
+use OCP\IGroupManager;
 use OCP\IL10N;
 use OCP\IURLGenerator;
+use OCP\IUser;
 use OCP\IUserManager;
 use Sabre\DAV\UUIDUtil;
+use Throwable;
 
 class AccountService {
 	/** @var IL10N */
@@ -42,12 +46,16 @@ class AccountService {
 	private $cfsslHandler;
 	/** @var FileMapper */
 	private $fileMapper;
+	/** @var ReportDao */
+	private $reportDao;
 	/** @var string */
 	private $pfxFilename = 'signature.pfx';
 	/** @var \OCA\Libresign\DbFile */
 	private $fileData;
 	/** @var \OCA\Files\Node\File */
 	private $fileToSign;
+	/** @var IGroupManager */
+	private $groupManager;
 
 	public function __construct(
 		IL10N $l10n,
@@ -56,10 +64,12 @@ class AccountService {
 		FolderService $folder,
 		IRootFolder $root,
 		FileMapper $fileMapper,
+		ReportDao $reportDao,
 		IConfig $config,
 		NewUserMailHelper $newUserMail,
 		IURLGenerator $urlGenerator,
-		CfsslHandler $cfsslHandler
+		CfsslHandler $cfsslHandler,
+		IGroupManager $groupManager
 	) {
 		$this->l10n = $l10n;
 		$this->fileUserMapper = $fileUserMapper;
@@ -67,10 +77,12 @@ class AccountService {
 		$this->folder = $folder;
 		$this->root = $root;
 		$this->fileMapper = $fileMapper;
+		$this->reportDao = $reportDao;
 		$this->config = $config;
 		$this->newUserMail = $newUserMail;
 		$this->urlGenerator = $urlGenerator;
 		$this->cfsslHandler = $cfsslHandler;
+		$this->groupManager = $groupManager;
 	}
 
 	public function validateCreateToSign(array $data) {
@@ -305,8 +317,11 @@ class AccountService {
 			case 'url':
 				$pdf = ['url' => $this->urlGenerator->linkToRoute('libresign.page.getPdfUser', ['uuid' => $uuid])];
 				break;
-			case 'file':
+			case 'nodeId':
 				$pdf = ['nodeId' => $fileToSign->getId()];
+				break;
+			case 'file':
+				$pdf = ['file' => $fileToSign];
 				break;
 		}
 		$return['sign'] = [
@@ -327,5 +342,68 @@ class AccountService {
 		} catch (\Throwable $th) {
 		}
 		return false;
+	}
+
+	/**
+	 * Get PDF node by UUID
+	 *
+	 * @param string $uuid
+	 * @throws Throwable
+	 * @return \OCP\Files\File
+	 */
+	public function getPdfByUuid(string $uuid): \OCP\Files\File {
+		$fileData = $this->fileMapper->getByUuid($uuid);
+		Filesystem::initMountPoints($fileData->getUserId());
+
+		$file = $this->root->getById($fileData->getNodeId())[0];
+		$filePath = $file->getPath();
+
+		$fileUser = $this->fileUserMapper->getByFileId($fileData->getId());
+		$signedUsers = array_filter($fileUser, function ($row) {
+			return !is_null($row->getSigned());
+		});
+		if (count($fileUser) === count($signedUsers)) {
+			$filePath = preg_replace(
+				'/' . $file->getExtension() . '$/',
+				$this->l10n->t('signed') . '.' . $file->getExtension(),
+				$filePath
+			);
+		}
+		// If signed, return signed file
+		if ($this->root->nodeExists($filePath)) {
+			/** @var \OCP\Files\File */
+			$file = $this->root->get($filePath);
+		}
+		return $file;
+	}
+
+	public function canRequestSign(?IUser $user = null): bool {
+		if (!$user) {
+			return false;
+		}
+		$authorized = json_decode($this->config->getAppValue(Application::APP_ID, 'webhook_authorized', '["admin"]'));
+		if (empty($authorized)) {
+			return false;
+		}
+		$userGroups = $this->groupManager->getUserGroupIds($user);
+		if (!array_intersect($userGroups, $authorized)) {
+			return false;
+		}
+		return true;
+	}
+
+	public function list(IUser $user, $page = null, $limit = 15) {
+		$return = $this->reportDao->getFilesAssociatedFilesWithMeFormatted($user->getUID(), $page, $limit);
+		return [
+			'data' => $return,
+			'pagination' => [
+				'total' => $this->reportDao->getTotalFilesAssociatedFilesWithMe($user->getUID()),
+				'current' => '',
+				'next' => '',
+				'prev' => '',
+				'last' => '',
+				'first' => ''
+			]
+		];
 	}
 }