+++ tests/Unit/Service/WebhookServiceTest.php
@@ -400,10 +400,12 @@ final class WebhookServiceTest extends \OCA\Libresign\Tests\Unit\TestCase {
 			->method('__call')
 			->withConsecutive(
 				[$this->equalTo('getUuid')],
+				[$this->equalTo('getNodeId')],
 				[$this->equalTo('getId')]
 			)
 			->will($this->returnValueMap([
 				['getUuid', [], 'uuid-here'],
+				['getNodeId', [], 123],
 				['getId', [], 123]
 			]));
 		$this->file->method('getByUuid')->will($this->returnValue($file));
@@ -578,6 +580,22 @@ final class WebhookServiceTest extends \OCA\Libresign\Tests\Unit\TestCase {
 		]);
 	}
 
+	public function testValidateUserWithNameAndInvalidEmail() {
+		$this->expectExceptionMessage('User %s needs an email address');
+
+		$this->service->validate([
+			'file' => ['base64' => 'dGVzdA=='],
+			'name' => 'test',
+			'users' => [
+				[
+					'email' => '',
+					'name' => 'User'
+				]
+			],
+			'userManager' => $this->user
+		]);
+	}
+
 	public function testValidateUserWithInvalidEmail() {
 		$this->expectExceptionMessage('Invalid email: user %s');
 
@@ -625,10 +643,34 @@ final class WebhookServiceTest extends \OCA\Libresign\Tests\Unit\TestCase {
 		$this->assertNull($actual);
 	}
 
-	public function testIndexWithoutPermission() {
+	public function testIndexWithoutUserManager() {
 		$this->expectExceptionMessage('You are not allowed to request signing');
 
 		$this->service->validate([]);
+	}
+
+	public function testIndexWithoutPermission() {
+		$this->expectExceptionMessage('You are not allowed to request signing');
+
+		$this->config = $this->createMock(IConfig::class);
+		$this->config
+			->method('getAppValue')
+			->willReturn('');
+		$this->service = new WebhookService(
+			$this->config,
+			$this->groupManager,
+			$this->l10n,
+			$this->file,
+			$this->fileUser,
+			$this->folder,
+			$this->clientService,
+			$this->userManager,
+			$this->mail,
+			$this->logger
+		);
+		$this->service->validate([
+			'userManager' => 'fake'
+		]);
 	}
 
 	public function testNotifyCallback() {