+++ CHANGELOG.md
@@ -20,13 +20,17 @@ Types of changes:
 - GitHub action to add a changelog reminder
 - List of documents
 - Test for validation of Swagger documentation
+- Added integration with Approval app. Thanks to @eneiluj
+- Endpoint to list LibreSign files
 
 # Changed
-- Increment of coverage
+- Increment of coverage on backend code
 - Bug fixes and refactorings resulting from increased coverage
+- Use name of user on error message when email is empty
 
 # Removed
 - Removed dsv folder
+- Removed docs folder
 
 # Fixed
 - Rendering files tab in Nextcloud 20 and 21